refactor: organize client versions feature

Copilot-Session: 26223ed6-096e-4060-8c97-01cb7d0ee3ab
