+++ src/components/features/client-versions/index.ts
@@ -1,0 +1,1 @@
+export { default as ClientVersionsView } from './ClientVersionsView';
